test : model패키지 객체 테스트 추가
